+++ Exercises/Cellterion/smartphone-service/src/main/java/com/cellterion/smartphoneservice/service/SmartphoneService.java
@@ -1,5 +1,6 @@
 package com.cellterion.smartphoneservice.service;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import com.cellterion.smartphoneservice.model.Smartphone;
@@ -16,6 +17,6 @@ public interface SmartphoneService {
 
 	Smartphone[] getSmartphonesByBrandName(String brandName);
 
-	List<Smartphone> getSmartphonesByUserBudget(double userBudget);
+	List<Smartphone> getSmartphonesByUserBudget(BigDecimal userBudget);
 
 }
